Solve conflicts, fix syntax
