+++ src/main/java/gov/milove/main/controller/impl/TextBannerControllerImpl.java
@@ -1,8 +1,9 @@
-package gov.milove.controllers;
+package gov.milove.main.controller.impl;
 
-import gov.milove.controllers.abstr.ILinkBannerController;
-import gov.milove.domain.LinkBanner;
-import gov.milove.repositories.jpa.LinkBannerRepository;
+import gov.milove.main.controller.TextBannerController;
+import gov.milove.main.domain.TextBanner;
+import gov.milove.main.repository.jpa.TextBannerRepository;
+import gov.milove.main.util.EntityMapper;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Example;
@@ -12,37 +13,37 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
-import static gov.milove.util.EntityMapper.map;
 import static org.springframework.http.HttpStatus.CONFLICT;
 import static org.springframework.http.HttpStatus.CREATED;
 
 @RestController
 @RequiredArgsConstructor
-public class LinkBannerController implements ILinkBannerController {
+public class TextBannerControllerImpl implements TextBannerController {
 
-    private final LinkBannerRepository repo;
+    private final TextBannerRepository repo;
 
     @Override
-    public List<LinkBanner> getAll() {
+    public List<TextBanner> getAll() {
         return repo.findAll(Sort.by("createdOn").descending());
     }
 
     @Override
-    public ResponseEntity<Long> createBanner(LinkBanner banner) {
+    public ResponseEntity<Long> createBanner(TextBanner banner) {
         if (repo.exists(Example.of(banner))) {
             return ResponseEntity.status(CONFLICT).build();
         } else {
-            LinkBanner saved = repo.save(banner);
+            TextBanner saved = repo.save(banner);
             return ResponseEntity.status(CREATED).body(saved.getId());
         }
     }
 
-    @Override
-    public ResponseEntity<?> update(LinkBanner banner) {
-        if (banner.getId() == null) return ResponseEntity.badRequest().build();
-        LinkBanner saved = repo.findById(banner.getId()).orElseThrow(EntityNotFoundException::new);
 
-        map(banner, saved).mapEmptyString(false).map();
+    @Override
+    public ResponseEntity<?> update(TextBanner banner) {
+        if (banner.getId() == null) return ResponseEntity.badRequest().build();
+        TextBanner saved = repo.findById(banner.getId()).orElseThrow(EntityNotFoundException::new);
+
+        EntityMapper.map(banner, saved).mapEmptyString(false).mapNull(false).map();
 
         repo.save(saved);
         return ResponseEntity.ok().build();
